Released slrn 0.9.7.4pl3
git-svn-id: https://slrn.svn.sourceforge.net/svnroot/slrn/trunk@143 b66dfc27-5e3b-0410-ac45-e82dab92a7d9

+++ src/post.c
@@ -1043,7 +1043,8 @@ static int insert_custom_header (char *fmt, FILE *fp) /*{{{*/
 	       return fputs (fmt, fp);
 	     else
 	       {
-		  if ((fwrite (fmt, 1, (unsigned int) (s - fmt), fp)) < s-fmt)
+		  if (fwrite (fmt, 1, (unsigned int) (s - fmt), fp) <
+		      (unsigned int) (s-fmt))
 		    return -1;
 		  fmt = s;
 		  continue;
